requirements.txt
altair==5.3.0
attrs==23.2.0
awscli==1.33.33
blinker==1.8.2
boto3==1.34.151
botocore==1.34.151
branca==0.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docker==7.1.0
docutils==0.16
duckdb==1.0.0
filelock==3.15.4
folium==0.17.0
fonttools==4.53.0
footballpitchplot==0.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.2.1
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
highlight-text==0.2
idna==3.7
intel-openmp==2021.4.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpld3==0.5.10
mplsoccer==1.2.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
plotly-express==0.4.1
plotly-football-pitch==0.0.3
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sql_metadata==2.12.0
sqlparse==0.5.1
st-gsheets-connection==0.1.0
statsmodels==0.14.2
streamlit==1.37.1
sympy==1.13.1
tableauserverclient==0.31
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.70
ultralytics-thop==2.0.0
urllib3==2.0.7
validators==0.33.0
voluptuous==0.15.2
watchdog==4.0.1
xyzservices==2024.6.0
yolo==0.3.1
haversine==2.8.1
shapely==2.0.6
gql==3.4.0
mysql-connector-python==8.1.0
